requirements.txt
# Compatible with Python 3.8–3.10

opencv-python==4.8.0.76
face_recognition==1.3.0
numpy==1.24.4
Pillow==9.5.0
mysql-connector-python==8.0.33
python-dotenv==1.0.1
